Move Painter into a separate file
